Delete nul
I am unable to pull this repo currently. I talked to someone else who also couldn't refresh it. I think it is related to the nul file, which is apparently not allowed for windows file paths. I don't know the history or purpose of this file, but I suspect removing it or renaming it will fix that issue.